requirements.txt
# Azure SDK
azure-identity>=1.12.0
azure-mgmt-costmanagement>=2.0.0
azure-mgmt-resource>=20.0.0
azure-mgmt-compute>=30.0.0
azure-keyvault-secrets>=4.6.0
azure-monitor-query>=1.0.0
azure-loganalytics>=1.0.0
azure-storage-blob>=12.16.0

# Infrastructure as Code
terraform>=1.5.0

# Python utilities
python-dateutil>=2.8.0

# Testing & Quality
pytest>=7.0.0
flake8>=6.0.0
black>=23.1.0
